+++ src/sns/commands/spec/values/proposals.rs
@@ -7,10 +7,14 @@
 use crate::sns::report::{SnsProposalStatusFilter, SnsProposalTopicFilter, SnsProposalsSort};
 use clap::ValueEnum;
 
-pub(in crate::sns::commands) const SNS_PROPOSALS_SORT_VALUE_NAME: &str =
-    "api|id|title|action|yes|no|total-votes|created|decided|executed|failed";
-pub(in crate::sns::commands) const SNS_PROPOSALS_LOCAL_SORT_VALUE_NAME: &str =
-    "id|title|action|yes|no|total-votes|created|decided|executed|failed";
+pub(in crate::sns::commands) const SNS_PROPOSALS_SORT_VALUE_NAME: &str = concat!(
+    "api|id|status|title|action|yes|no|total-votes|",
+    "ballots|reject-cost|created|decided|executed|failed"
+);
+pub(in crate::sns::commands) const SNS_PROPOSALS_LOCAL_SORT_VALUE_NAME: &str = concat!(
+    "id|status|title|action|yes|no|total-votes|",
+    "ballots|reject-cost|created|decided|executed|failed"
+);
 
 ///
 /// SnsProposalsSortArg
@@ -23,11 +27,14 @@ pub(in crate::sns::commands) enum SnsProposalsSortArg {
     #[default]
     Api,
     Id,
+    Status,
     Title,
     Action,
     Yes,
     No,
     TotalVotes,
+    Ballots,
+    RejectCost,
     Created,
     Decided,
     Executed,
@@ -39,11 +46,14 @@ impl From<SnsProposalsSortArg> for SnsProposalsSort {
         match value {
             SnsProposalsSortArg::Api => Self::Api,
             SnsProposalsSortArg::Id => Self::Id,
+            SnsProposalsSortArg::Status => Self::Status,
             SnsProposalsSortArg::Title => Self::Title,
             SnsProposalsSortArg::Action => Self::Action,
             SnsProposalsSortArg::Yes => Self::Yes,
             SnsProposalsSortArg::No => Self::No,
             SnsProposalsSortArg::TotalVotes => Self::TotalVotes,
+            SnsProposalsSortArg::Ballots => Self::Ballots,
+            SnsProposalsSortArg::RejectCost => Self::RejectCost,
             SnsProposalsSortArg::Created => Self::Created,
             SnsProposalsSortArg::Decided => Self::Decided,
             SnsProposalsSortArg::Executed => Self::Executed,